+++ src/components/fender.vue
@@ -1,6 +1,6 @@
 <template>
     <div>
-        Register
+        fender
     </div>
 </template>
 <script>
