+++ src/components/gibson.vue
@@ -20,7 +20,7 @@
 	</div>
 -->
     <div>
-        Login
+        gibson
     </div>
 </template>
 <script>
